renamed receipt -> add retail receipt
